# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quwan Chess API (Contract-first v1)
  version: 1.0.0
  description: |
    Contract-first API for Quwan Chess.

    Notes:
    - Planned public endpoints use the prefix /api/v1/... with a unified envelope { code, message, data }.
    - Current dev endpoints implemented in this repo are /user/register and /user/login (no envelope). They are included as legacy for compatibility and will be deprecated once /api/v1/... is adopted.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register via phone + password
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthTokens"
              examples:
                ok:
                  value:
                    code: 0
                    message: 注册成功
                    data:
                      accessToken: eyJhbGciOi...
                      refreshToken: eyJhbGciOi...
                      expiresIn: 1800
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                invalidPhone:
                  value:
                    code: 1001
                    message: 手机号格式不正确
                    data: null
                phoneTaken:
                  value:
                    code: 1002
                    message: 手机号已被注册
                    data: null
                weakPassword:
                  value:
                    code: 1003
                    message: 密码过于简单
                    data: null
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login via phone + password
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthTokens"
              examples:
                ok:
                  value:
                    code: 0
                    message: 登录成功
                    data:
                      accessToken: eyJhbGciOi...
                      refreshToken: eyJhbGciOi...
                      expiresIn: 1800
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                incorrectCredentials:
                  value:
                    code: 401
                    message: 账号或密码错误
                    data: null
                userNotFound:
                  value:
                    code: 401
                    message: 用户不存在或已被禁用
                    data: null
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh and rotate tokens
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRefreshRequest"
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthTokens"
              examples:
                ok:
                  value:
                    code: 0
                    message: 刷新成功
                    data:
                      accessToken: eyJhbGciOi...
                      refreshToken: eyJhbGciOi...
                      expiresIn: 1800
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                invalidRefresh:
                  value:
                    code: 401
                    message: 刷新令牌无效
                    data: null
                expiredRefresh:
                  value:
                    code: 401
                    message: 刷新令牌已过期
                    data: null
                revoked:
                  value:
                    code: 401
                    message: 刷新令牌已注销
                    data: null
...
  /api/v1/auth/sms:
    post:
      tags: [Auth]
      summary: Send SMS code
      operationId: authSms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsRequest"
      responses:
        "200":
          description: SMS sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseSmsSent"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                invalidPhone:
                  value:
                    code: 1001
                    message: 手机号格式不正确
                    data: null
                rateLimited:
                  value:
                    code: 1004
                    message: 请求过于频繁，请稍后再试
                    data: null
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user
      operationId: authLogout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOk"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                unauthorized:
                  value:
                    code: 401
                    message: 未认证或令牌无效
                    data: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseAuthTokens:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: "#/components/schemas/AuthTokens"
      required: [code, message, data]
...
    ApiResponseError:
      type: object
      properties:
        code:
          type: integer
          example: 1001
        message:
          type: string
          example: 参数错误
        data:
          oneOf:
            - type: "null"
            - type: object
              description: 参数校验错误详情
              properties:
                details:
                  type: array
                  items: { type: string }
      required: [code, message]
    AuthRegisterRequest:
      type: object
      required: [type, phone, password]
      properties:
        type:
          type: string
          enum: [phone]
          default: phone
        phone:
          type: string
          example: "13800000000"
        code:
          type: string
          description: SMS verification code (optional if disabled)
          example: "8523"
        password:
          type: string
          format: password
          minLength: 6
          example: Abc12345
        email:
          type: string
          format: email
          nullable: true
    AuthLoginRequest:
      type: object
      required: [type, phone, password]
      properties:
        type:
          type: string
          enum: [phone]
          default: phone
        phone:
          type: string
          example: "13800000000"
        password:
          type: string
          format: password
          example: Abc12345
    AuthRefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOi...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: eyJhbGciOi...
        refreshToken:
          type: string
          description: JWT refresh token
          example: eyJhbGciOi...
        expiresIn:
          type: integer
          description: Access token TTL in seconds
          example: 1800
      required: [accessToken, refreshToken]
...
    SmsRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: "13800000000" }
    ApiResponseSmsSent:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: 短信已发送 }
        data:
          type: object
          properties:
            requestId: { type: string, example: sms_9f3a2 }
            expireIn: { type: integer, example: 300 }
      required: [code, message]
    ApiResponseOk:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true }
...